Moved oled codes to keyboard level & changed some functions
